Move asset hooks from modal to container
