+++ src/components/molecules/Common/AssetModal/AssetContainer/index.tsx
@@ -46,7 +46,7 @@ export type Props = {
   onRemove?: (assetIds: string[]) => void;
   onGetMore?: () => void;
   selectAssetUrl?: (asset?: string) => void;
-  onSelect?: (asset?: Asset | undefined) => void;
+  onSelect?: (asset?: Asset) => void;
   onSortChange?: (type?: string, reverse?: boolean) => void;
   onSearch?: (term?: string) => void;
   handleShowURL?: (assets?: Asset[]) => void;
